Move tests into TestCreateActions
